Add hook type
Hooks in Cucumber can either execute before/after a scenario or
before/after a step. Depending on which implementation is used these are
put into the json report as either all part of the `element.steps`
field[1] or the `element.before` and `element.steps` fields[2][3].

So to create a json reports from messages the type of hook must be
included in the message.

 1. https://github.com/cucumber/cucumber-json-converter/blob/main/testdata/cucumber-js/7.3.2/json/failing-before-hook.json
 2. https://github.com/cucumber/cucumber-json-converter/blob/main/testdata/cucumber-jvm/7.2.3/json/failing-before-hook.json
 3. https://github.com/cucumber/cucumber-json-converter/blob/main/testdata/cucumber-jvm/7.2.3/json/failing-after-step-hook.json