Include description in documentation (#321)
Updated the markdown template to include the object and field
descriptions of all messages. This makes it a bit easier to understand
what the messages in the protocol are about.

Because descriptions contain multi-line strings this doesn't work well
with the table format. But we can replicate the effect using headers at
different levels.

Note: Every property is prefixed with a `.` to ensure message names are
unique and Markdown links keep working.