Add NdjsonToMessageReader
Unlike the `NdjsonToMessageIterable` this  implementation is symetric
with the `MessageToNdjsonWriter`. And like other readers it supports
reading the contents as a stream. Streams are in general a bit easier to
work with.

Addtionally the `MessageToNdjsonWriter` is switched to a
`BufferedWriter` and no longer flushes immediately. This should improve
IO performance by allowing larger chunks to be read.